add reactions onboarding popover
