+++ packages/atlas/src/components/_video/ReactionsOnboardingPopover/ReactionsOnboardingPopover.styles.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled'
+
+import { cVar, sizes } from '@/styles'
+
+export const PopoverIllustrationWrapper = styled.div`
+  background-color: ${cVar('colorBackground')};
+`
+export const PopoverContentWrapper = styled.div`
+  padding: ${sizes(4)};
+`
