+++ packages/atlas/src/components/_video/ReactionsOnboardingPopover/index.ts
@@ -1,0 +1,1 @@
+export * from './ReactionsOnboardingPopover'
